Show release dates in collection grid, move views to root module
- CollectionView shows Release Date column from birthday data
- Collection grid takes full page height
- BirthdayView and CollectionView moved to root module (app.stolat)
  to avoid modulith cycles — cross-module views belong in root
- BirthdayService gains findReleaseDatesByMusicBrainzId for view lookup
