+++ package.json
@@ -1,0 +1,147 @@
+{
+  "name": "no-name",
+  "license": "UNLICENSED",
+  "type": "module",
+  "dependencies": {
+    "@vaadin/aura": "25.0.4",
+    "@vaadin/common-frontend": "0.0.19",
+    "@vaadin/react-components": "25.0.4",
+    "@vaadin/vaadin-development-mode-detector": "2.0.7",
+    "@vaadin/vaadin-lumo-styles": "25.0.4",
+    "@vaadin/vaadin-themable-mixin": "25.0.4",
+    "@vaadin/vaadin-usage-statistics": "2.1.3",
+    "date-fns": "4.1.0",
+    "lit": "3.3.2",
+    "react": "19.2.4",
+    "react-dom": "19.2.4",
+    "react-router": "7.12.0"
+  },
+  "devDependencies": {
+    "@babel/preset-react": "7.28.5",
+    "@preact/signals-react-transform": "0.6.0",
+    "@rollup/plugin-replace": "6.0.3",
+    "@rollup/pluginutils": "5.3.0",
+    "@types/node": "25.0.10",
+    "@types/react": "19.2.9",
+    "@types/react-dom": "19.2.3",
+    "@vitejs/plugin-react": "5.1.2",
+    "magic-string": "0.30.21",
+    "rollup-plugin-brotli": "3.1.0",
+    "rollup-plugin-visualizer": "6.0.5",
+    "strip-css-comments": "5.0.0",
+    "transform-ast": "2.4.4",
+    "typescript": "5.9.3",
+    "vite": "7.3.1",
+    "vite-plugin-checker": "0.12.0",
+    "workbox-build": "7.4.0"
+  },
+  "vaadin": {
+    "dependencies": {
+      "@vaadin/aura": "25.0.4",
+      "@vaadin/common-frontend": "0.0.19",
+      "@vaadin/react-components": "25.0.4",
+      "@vaadin/vaadin-development-mode-detector": "2.0.7",
+      "@vaadin/vaadin-lumo-styles": "25.0.4",
+      "@vaadin/vaadin-themable-mixin": "25.0.4",
+      "@vaadin/vaadin-usage-statistics": "2.1.3",
+      "date-fns": "4.1.0",
+      "lit": "3.3.2",
+      "react": "19.2.4",
+      "react-dom": "19.2.4",
+      "react-router": "7.12.0"
+    },
+    "devDependencies": {
+      "@babel/preset-react": "7.28.5",
+      "@preact/signals-react-transform": "0.6.0",
+      "@rollup/plugin-replace": "6.0.3",
+      "@rollup/pluginutils": "5.3.0",
+      "@types/node": "25.0.10",
+      "@types/react": "19.2.9",
+      "@types/react-dom": "19.2.3",
+      "@vitejs/plugin-react": "5.1.2",
+      "magic-string": "0.30.21",
+      "rollup-plugin-brotli": "3.1.0",
+      "rollup-plugin-visualizer": "6.0.5",
+      "strip-css-comments": "5.0.0",
+      "transform-ast": "2.4.4",
+      "typescript": "5.9.3",
+      "vite": "7.3.1",
+      "vite-plugin-checker": "0.12.0",
+      "workbox-build": "7.4.0"
+    },
+    "hash": "bdce75e7f71cc66f34872c14c4944086207db2d057cf70fd9148d00ff2415fb8"
+  },
+  "overrides": {
+    "@vaadin/vaadin-development-mode-detector": "$@vaadin/vaadin-development-mode-detector",
+    "@vaadin/vaadin-usage-statistics": "$@vaadin/vaadin-usage-statistics",
+    "@vaadin/react-components": "$@vaadin/react-components",
+    "@vaadin/common-frontend": "$@vaadin/common-frontend",
+    "react-dom": "$react-dom",
+    "lit": "$lit",
+    "react": "$react",
+    "react-router": "$react-router",
+    "date-fns": "$date-fns",
+    "@vaadin/vaadin-themable-mixin": "$@vaadin/vaadin-themable-mixin",
+    "@vaadin/vaadin-lumo-styles": "$@vaadin/vaadin-lumo-styles",
+    "@vaadin/aura": "$@vaadin/aura",
+    "@vaadin/a11y-base": "25.0.4",
+    "@vaadin/accordion": "25.0.4",
+    "@vaadin/app-layout": "25.0.4",
+    "@vaadin/avatar": "25.0.4",
+    "@vaadin/avatar-group": "25.0.4",
+    "@vaadin/button": "25.0.4",
+    "@vaadin/card": "25.0.4",
+    "@vaadin/checkbox": "25.0.4",
+    "@vaadin/checkbox-group": "25.0.4",
+    "@vaadin/combo-box": "25.0.4",
+    "@vaadin/component-base": "25.0.4",
+    "@vaadin/confirm-dialog": "25.0.4",
+    "@vaadin/context-menu": "25.0.4",
+    "@vaadin/custom-field": "25.0.4",
+    "@vaadin/date-picker": "25.0.4",
+    "@vaadin/date-time-picker": "25.0.4",
+    "@vaadin/details": "25.0.4",
+    "@vaadin/dialog": "25.0.4",
+    "@vaadin/email-field": "25.0.4",
+    "@vaadin/field-base": "25.0.4",
+    "@vaadin/field-highlighter": "25.0.4",
+    "@vaadin/form-layout": "25.0.4",
+    "@vaadin/grid": "25.0.4",
+    "@vaadin/horizontal-layout": "25.0.4",
+    "@vaadin/icon": "25.0.4",
+    "@vaadin/icons": "25.0.4",
+    "@vaadin/input-container": "25.0.4",
+    "@vaadin/integer-field": "25.0.4",
+    "@vaadin/item": "25.0.4",
+    "@vaadin/list-box": "25.0.4",
+    "@vaadin/lit-renderer": "25.0.4",
+    "@vaadin/login": "25.0.4",
+    "@vaadin/markdown": "25.0.4",
+    "@vaadin/master-detail-layout": "25.0.4",
+    "@vaadin/menu-bar": "25.0.4",
+    "@vaadin/message-input": "25.0.4",
+    "@vaadin/message-list": "25.0.4",
+    "@vaadin/multi-select-combo-box": "25.0.4",
+    "@vaadin/notification": "25.0.4",
+    "@vaadin/number-field": "25.0.4",
+    "@vaadin/overlay": "25.0.4",
+    "@vaadin/password-field": "25.0.4",
+    "@vaadin/popover": "25.0.4",
+    "@vaadin/progress-bar": "25.0.4",
+    "@vaadin/radio-group": "25.0.4",
+    "@vaadin/scroller": "25.0.4",
+    "@vaadin/select": "25.0.4",
+    "@vaadin/side-nav": "25.0.4",
+    "@vaadin/split-layout": "25.0.4",
+    "@vaadin/tabs": "25.0.4",
+    "@vaadin/tabsheet": "25.0.4",
+    "@vaadin/text-area": "25.0.4",
+    "@vaadin/text-field": "25.0.4",
+    "@vaadin/time-picker": "25.0.4",
+    "@vaadin/tooltip": "25.0.4",
+    "@vaadin/upload": "25.0.4",
+    "@vaadin/router": "2.0.1",
+    "@vaadin/vertical-layout": "25.0.4",
+    "@vaadin/virtual-list": "25.0.4"
+  }
+}
